+++ app.css
@@ -30,12 +30,6 @@ textarea {
   border: 1px solid rgb(175, 175, 175);
 }
 
-input:focus,
-textarea:focus {
-  outline: none;
-  background: #fff5f9;
-}
-
 button {
   font: inherit;
   background: #ff0062;
